FAT: outsourced FAT implementations into seperate files
